requirements.txt
pandas
transformers
torch
pinecone-client
python-dotenv
nltk
tqdm
Flask
numpy
scikit-learn
